+++ src/resources/meta.ts
@@ -1,8 +1,11 @@
 import type { CsfloatHttpClient } from "../client.js";
 import type {
+  CsfloatAppMetaResponse,
   CsfloatExchangeRatesResponse,
   CsfloatLocationResponse,
   CsfloatNotaryMetaResponse,
+  CsfloatSchemaBrowseResponse,
+  CsfloatSchemaBrowseType,
   CsfloatSchemaResponse,
 } from "../types.js";
 
@@ -17,8 +20,16 @@ export class MetaResource {
     return this.client.get<CsfloatExchangeRatesResponse>("meta/exchange-rates");
   }
 
+  getApp(): Promise<CsfloatAppMetaResponse> {
+    return this.client.get<CsfloatAppMetaResponse>("meta/app");
+  }
+
   getLocation(): Promise<CsfloatLocationResponse> {
     return this.client.get<CsfloatLocationResponse>("meta/location");
+  }
+
+  getSchemaBrowse(type: CsfloatSchemaBrowseType): Promise<CsfloatSchemaBrowseResponse> {
+    return this.client.get<CsfloatSchemaBrowseResponse>("schema/browse", { type });
   }
 
   getNotary(): Promise<CsfloatNotaryMetaResponse> {